+++ src/main/java/com/barlinc/unusual_prehistory/UnusualPrehistory2Tab.java
@@ -59,6 +59,7 @@ public class UnusualPrehistory2Tab {
 
                         // Ambient
                         output.accept(DELITZSCHALA_SPAWN_EGG.get());
+                        output.accept(SETAPEDITES_SPAWN_EGG.get());
                         output.accept(ZHANGSOLVA_SPAWN_EGG.get());
 
                         // Misc