+++ src/main/java/com/barlinc/unusual_prehistory/client/models/entity/mob/update_6/ambient/SetapeditesModel.java
@@ -1,0 +1,46 @@
+package com.barlinc.unusual_prehistory.client.models.entity.mob.update_6.ambient;
+
+import com.barlinc.unusual_prehistory.client.models.entity.UP2Model;
+import com.barlinc.unusual_prehistory.entity.mob.update_6.ambient.Setapedites;
+import net.minecraft.client.model.geom.ModelPart;
+import net.minecraft.client.model.geom.PartPose;
+import net.minecraft.client.model.geom.builders.*;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
+import org.jetbrains.annotations.NotNull;
+
+@OnlyIn(Dist.CLIENT)
+@SuppressWarnings("FieldCanBeLocal, unused")
+public class SetapeditesModel extends UP2Model<Setapedites> {
+
+    private final ModelPart root;
+
+	public SetapeditesModel(ModelPart root) {
+        super(1.0F, 0);
+        this.root = root.getChild("root");
+	}
+
+	public static LayerDefinition createBodyLayer() {
+        MeshDefinition meshdefinition = new MeshDefinition();
+        PartDefinition partdefinition = meshdefinition.getRoot();
+
+        PartDefinition root = partdefinition.addOrReplaceChild("root", CubeListBuilder.create().texOffs(0, 6).addBox(-2.5F, -3.0F, -3.0F, 5.0F, 3.0F, 6.0F, new CubeDeformation(0.0F))
+                .texOffs(-3, 15).addBox(-1.5F, 0.0F, 3.0F, 3.0F, 0.0F, 7.0F, new CubeDeformation(0.0F))
+                .texOffs(0, 0).addBox(-3.5F, 0.0F, -4.0F, 7.0F, 0.0F, 6.0F, new CubeDeformation(0.0F)), PartPose.offset(0.0F, 24.0F, 0.0F));
+
+        return LayerDefinition.create(meshdefinition, 32, 32);
+	}
+
+	@Override
+	public void setupAnim(@NotNull Setapedites entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
+		this.root().getAllParts().forEach(ModelPart::resetPose);
+        if (entity.isInWaterOrBubble()) {
+            this.root.xRot = headPitch * ((float) Math.PI / 180F);
+        }
+	}
+
+	@Override
+	public @NotNull ModelPart root() {
+		return this.root;
+	}
+}