+++ src/main/java/com/barlinc/unusual_prehistory/registry/UP2ModelLayers.java
@@ -73,6 +73,7 @@ public class UP2ModelLayers {
     public static final ModelLayerLocation HYNERPETON = mainLayer("hynerpeton");
     public static final ModelLayerLocation MAMMOTH = mainLayer("mammoth");
     public static final ModelLayerLocation PROGNATHODON = mainLayer("prognathodon");
+    public static final ModelLayerLocation SETAPEDITES = mainLayer("setapedites");
 
     // Future
     public static final ModelLayerLocation BARINASUCHUS = mainLayer("barinasuchus");